Workshop 3

+++ routes/campsiteRouter.js
@@ -17,7 +17,7 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser, (req, res, next) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     //for handling http post requests
     Campsite.create(req.body)
       .then((campsite) => {
@@ -28,19 +28,23 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .put(authenticate.verifyUser, (req, res) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res) => {
     res.statusCode = 403;
     res.end("PUT operation not supported on '/campsites'.");
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.deleteMany()
-      .then((response) => {
-        res.statusCode = 200;
-        res.setHeader("Content-Type", "application/json");
-        res.json(response);
-      })
-      .catch((err) => next(err));
-  });
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.deleteMany()
+        .then((response) => {
+          res.statusCode = 200;
+          res.setHeader("Content-Type", "application/json");
+          res.json(response);
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 //The following is for routes of particular campsites
 campsiteRouter
@@ -55,13 +59,13 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser, (req, res) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res) => {
     res.statusCode = 403;
     res.end(
       `Post operation not supported on "/campsites/${req.params.campsiteId}."`
     );
   })
-  .put(authenticate.verifyUser, (req, res, next) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Campsite.findByIdAndUpdate(
       req.params.campsiteId,
       { $set: req.body },
@@ -74,15 +78,19 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.findByIdAndDelete(req.params.campsiteId)
-      .then((response) => {
-        res.statusCode = 200;
-        res.setHeader("Content-Type", "application/json");
-        res.json(response);
-      })
-      .catch((err) => next(err));
-  });
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.findByIdAndDelete(req.params.campsiteId)
+        .then((response) => {
+          res.statusCode = 200;
+          res.setHeader("Content-Type", "application/json");
+          res.json(response);
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 //route for campsites/:campsiteId/comments
 campsiteRouter
@@ -106,7 +114,7 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser, (req, res, next) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     //for handling http post requests to comments
     Campsite.findById(req.params.campsiteId)
       .then((campsite) => {
@@ -131,37 +139,41 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .put(authenticate.verifyUser, (req, res) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res) => {
     res.statusCode = 403;
     res.end(
       `PUT operation not supported on /campsites/${req.params.campsiteId}/comments.`
     );
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.findById(req.params.campsiteId)
-      .then((campsite) => {
-        if (campsite) {
-          for (let i = campsite.comments.length - 1; i >= 0; i--) {
-            campsite.comments.id(campsite.comments[i]._id).deleteOne();
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.findById(req.params.campsiteId)
+        .then((campsite) => {
+          if (campsite) {
+            for (let i = campsite.comments.length - 1; i >= 0; i--) {
+              campsite.comments.id(campsite.comments[i]._id).deleteOne();
+            }
+            campsite
+              .save()
+              .then((campsite) => {
+                res.statusCode = 200;
+                res.setHeader("Content-Type", "application/json");
+                res.json(campsite);
+              })
+              .catch((err) => next(err));
+          } else {
+            const err = new Error(
+              `Campsite ${req.params.campsiteId} was not found.`
+            );
+            err.status = 404;
+            return next(err);
           }
-          campsite
-            .save()
-            .then((campsite) => {
-              res.statusCode = 200;
-              res.setHeader("Content-Type", "application/json");
-              res.json(campsite);
-            })
-            .catch((err) => next(err));
-        } else {
-          const err = new Error(
-            `Campsite ${req.params.campsiteId} was not found.`
-          );
-          err.status = 404;
-          return next(err);
-        }
-      })
-      .catch((err) => next(err));
-  });
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 //endpoints for /:campsiteId/comments/:commentId
 campsiteRouter
@@ -191,31 +203,44 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser, (req, res) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res) => {
     //for handling http post requests to comments
     res.statusCode = 403;
     res.end(
       `Post operation not supported on /campsites/${req.params.campsiteId}/comments/${req.params.commentId}.`
     );
   })
-  .put(authenticate.verifyUser, (req, res, next) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
       .then((campsite) => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
-          if (req.body.rating) {
-            campsite.comments.id(req.params.commentId).rating = req.body.rating;
+          if (
+            campsite.comments
+              .id(req.params.commentId)
+              .author._id.equals(req.user._id)
+          ) {
+            if (req.body.rating) {
+              campsite.comments.id(req.params.commentId).rating =
+                req.body.rating;
+            }
+            if (req.body.text) {
+              campsite.comments.id(req.params.commentId).text = req.body.text;
+            }
+            campsite
+              .save()
+              .then((campsite) => {
+                res.statusCode = 200;
+                res.setHeader("Content-Type", "application/json");
+                res.json(campsite);
+              })
+              .catch((err) => next(err));
+          } else {
+            const err = new Error(
+              "Your are not authorized to update this comment"
+            );
+            err.status = 403;
+            return next(err);
           }
-          if (req.body.text) {
-            campsite.comments.id(req.params.commentId).text = req.body.text;
-          }
-          campsite
-            .save()
-            .then((campsite) => {
-              res.statusCode = 200;
-              res.setHeader("Content-Type", "application/json");
-              res.json(campsite);
-            })
-            .catch((err) => next(err));
         } else if (!campsite) {
           const err = new Error(
             `Campsite ${req.params.campsiteId} was not found.`
@@ -232,34 +257,44 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.findById(req.params.campsiteId)
-      .then((campsite) => {
-        if (campsite && campsite.comments.id(req.params.commentId)) {
-          campsite.comments.id(req.params.commentId).deleteOne();
-          campsite
-            .save()
-            .then((campsite) => {
-              res.statusCode = 200;
-              res.setHeader("Content-Type", "application/json");
-              res.json(campsite);
-            })
-            .catch((err) => next(err));
-        } else if (!campsite) {
-          const err = new Error(
-            `Campsite ${req.params.campsiteId} was not found.`
-          );
-          err.status = 404;
-          return next(err);
-        } else {
-          const err = new Error(
-            `Comment ${req.params.commentId} was not found.`
-          );
-          err.status = 404;
-          return next(err);
-        }
-      })
-      .catch((err) => next(err));
-  });
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.findById(req.params.campsiteId)
+        .then((campsite) => {
+          if (campsite && campsite.comments.id(req.params.commentId)) {
+            if (
+              campsite.comments
+                .id(req.params.commentId)
+                .author._id.equals(req.user._id)
+            ) {
+              campsite.comments.id(req.params.commentId).deleteOne();
+              campsite
+                .save()
+                .then((campsite) => {
+                  res.statusCode = 200;
+                  res.setHeader("Content-Type", "application/json");
+                  res.json(campsite);
+                })
+                .catch((err) => next(err));
+            } else if (!campsite) {
+              const err = new Error(
+                `Campsite ${req.params.campsiteId} was not found.`
+              );
+              err.status = 404;
+              return next(err);
+            } else {
+              const err = new Error(
+                `Comment ${req.params.commentId} was not found.`
+              );
+              err.status = 404;
+              return next(err);
+            }
+          }
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 module.exports = campsiteRouter;
